feat: add workflow to label issues with "Needs verification" when moved to Done
This workflow is triggered when an issue is moved to the "Done" column in the project board. It automatically adds the "Needs verification" label to the issue. The label will be used to indicate that the issue needs to be verified in the staging deployment before it can be considered fully resolved. Once the verification is complete, the label can be manually removed by the person validating the issue.